Delete the QuranAPI class and sp the api functions 殺
