+++ src/main/java/com/anas/alqurancloudapi/Mapper.java
@@ -1,0 +1,10 @@
+package com.anas.alqurancloudapi;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import java.io.Serializable;
+
+public interface Mapper extends Serializable {
+     ObjectMapper mapper = new ObjectMapper();
+
+}
