+++ src/main/java/com/anas/alqurancloudapi/quran/Ayah.java
@@ -1,14 +1,18 @@
 package com.anas.alqurancloudapi.quran;
 
+import com.anas.alqurancloudapi.Mapper;
+import com.anas.alqurancloudapi.api.Requester;
+import com.anas.alqurancloudapi.consts.Constants;
+import com.anas.alqurancloudapi.consts.Surahs;
 import com.anas.alqurancloudapi.quran.edition.Edition;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.io.Serializable;
+import java.io.IOException;
 import java.util.Arrays;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Ayah implements Serializable {
+public class Ayah implements Mapper {
     private final short number;
     private final String text;
     @JsonProperty("audio")
@@ -162,6 +166,157 @@ public class Ayah implements Serializable {
      */
     public boolean isSajdah() {
         return sajdah;
+    }
+
+    // api
+    /**
+     * It takes an ayah number and an edition (optional) and returns an Ayah object
+     *
+     * @param ayahNumber The number of the ayah, numbered from 1 to 6348.
+     * @param edition    The edition of the Quran that you want to get the ayah from.
+     * @return An Ayah object.
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getAyah(final int ayahNumber, final Edition edition) throws IOException {
+        // Checking if the surah number is valid.
+        if (ayahNumber < 1 || ayahNumber > Constants.AYAHS_COUNT) {
+            throw new IllegalArgumentException("Ayah number must be between 1 and " + Constants.AYAHS_COUNT);
+        }
+        final var inputStream = Requester.sendRequest("ayah/" + ayahNumber
+                + (edition != null ? "/" + edition.getIdentifier() : ""));
+        return mapper.readValue(inputStream, Ayah.class);   // Ayah object
+    }
+
+    /**
+     * This function returns an Ayah object for the given ayah number and edition
+     *
+     * @param ayahNumber        The ayah number you want to get, numbered from 1 to 6348.
+     * @param editionIdentifier The identifier of the edition you want to use.
+     * @return An Ayah object.
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getAyah(final int ayahNumber, final String editionIdentifier) throws IOException {
+        return getAyah(ayahNumber, new Edition(editionIdentifier));
+    }
+
+    /**
+     * This function returns an Ayah object for the given ayah number
+     *
+     * @param ayahNumber The ayah number you want to get, numbered from 1 to 6348.
+     * @return An Ayah object
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getAyah(final int ayahNumber) throws IOException {
+        return getAyah(ayahNumber, (Edition) null);
+    }
+
+    /**
+     * Get a random ayah from the Quran
+     *
+     * @param edition The edition of the Quran you want to get the ayah from.
+     * @return A random ayah from the Quran.
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getRandomAyah(final Edition edition) throws IOException {
+        return getAyah((int) (Math.random() * Constants.AYAHS_COUNT), edition);
+    }
+
+    /**
+     * This function returns a random Ayah from the Quran
+     *
+     * @param editionIdentifier The identifier of the edition you want to use.
+     * @return A random ayah from the Quran.
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getRandomAyah(final String editionIdentifier) throws IOException {
+        return getRandomAyah(new Edition(editionIdentifier));
+    }
+
+    /**
+     * This function returns a random Ayah from the Quran
+     *
+     * @return A random ayah from the Quran.
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getRandomAyah() throws IOException {
+        return getRandomAyah((Edition) null);
+    }
+
+    /**
+     * This function returns an ayah from specified surah and ayah number
+     *
+     * @param surah      The surah you want to get the ayah from.
+     * @param ayahNumber The ayah number in the surah.
+     * @param edition    The edition of the Quran you want to use, (null mens default edition).
+     * @return An Ayah object.
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getAyahFromSurah(final Surahs surah,
+                                        final int ayahNumber,
+                                        final Edition edition) throws IOException {
+        return getAyah(surah.getFactAyahNumber(ayahNumber), edition);
+    }
+
+    /**
+     * This function returns an Ayah object from a Surah object, ayah number, and edition identifier
+     *
+     * @param surah             The surah you want to get the ayah from.
+     * @param ayahNumber        The ayah number you want to get.
+     * @param editionIdentifier The identifier of the edition you want to use.
+     * @return An Ayah object.
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getAyahFromSurah(final Surahs surah,
+                                        final int ayahNumber,
+                                        final String editionIdentifier) throws IOException {
+        return getAyahFromSurah(surah, ayahNumber, new Edition(editionIdentifier));
+    }
+
+    /**
+     * Get the Ayah object from specified surah.
+     *
+     * @param surah      The surah you want to get the ayah from.
+     * @param ayahNumber The ayah number in the surah
+     * @return An Ayah object
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getAyahFromSurah(final Surahs surah, final int ayahNumber) throws IOException {
+        return getAyah(surah.getFactAyahNumber(ayahNumber), (Edition) null);
+    }
+
+    /**
+     * This function returns an Ayah object from a specified surah and edition
+     *
+     * @param surah   The surah you want to get the ayah from.
+     * @param edition The edition of the Quran you want to use.
+     * @return A random ayah from a surah.
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getRandomAyahFromSurah(final Surahs surah, final Edition edition) throws IOException {
+        return getAyah(surah.getFactAyahNumber((int) (Math.random() * surah.getNumberOfAyahs())), edition);
+    }
+
+    /**
+     * This function returns an Ayah object from a specified surah and edition.
+     *
+     * @param surah             The surah you want to get a random ayah from.
+     * @param editionIdentifier The identifier of the edition you want to use.
+     * @return An Ayah object
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getRandomAyahFromSurah(final Surahs surah, final String editionIdentifier) throws IOException {
+        return getRandomAyahFromSurah(surah, new Edition(editionIdentifier));
+    }
+
+    /**
+     * This function returns a random ayah from a given surah
+     *
+     * @param surah The surah you want to get a random ayah from.
+     * @return An Ayah object
+     * @throws IOException If an error occurs while communicating with the API.
+     */
+    public static Ayah getRandomAyahFromSurah(final Surahs surah) throws IOException {
+        return getRandomAyahFromSurah(surah, (Edition) null);
     }
 
     @Override
