+++ src/main/java/com/anas/alqurancloudapi/search/SearchResult.java
@@ -1,0 +1,21 @@
+package com.anas.alqurancloudapi.search;
+
+import com.anas.alqurancloudapi.quran.Ayah;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+/**
+ * @author: <a href="https://github.com/anas-elgarhy">Anas Elgarhy</a>
+ * @date: 7/26/22
+ */
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class SearchResult {
+    private final Ayah[] matches;
+
+    public SearchResult() {
+        matches = null;
+    }
+
+    public Ayah[] getMatches() {
+        return matches;
+    }
+}
